requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
PyYAML==3.13
pandas==0.24.2
numpy==1.16.2
findspark==1.3.0
astral==1.10.1


